+++ src/main/java/lv/oug/android/domain/DatabaseHelper.java
@@ -17,11 +17,12 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static final int DATABASE_VERSION = 1;
     private static final int NOT_MAPPED = -1;
     private static ClassLogger logger = new ClassLogger(DatabaseHelper.class);
-    private Dao<Event, Integer> eventDao;
-    private Dao<Article, Integer> articleDao;
-    private Dao<Sponsor, Integer> sponsorDao;
-    private Dao<Material, Integer> materialDao;
-    private Dao<Contact, Integer> contactDao;
+
+    private Dao<Event, Long> eventDao;
+    private Dao<Article, Long> articleDao;
+    private Dao<Sponsor, Long> sponsorDao;
+    private Dao<Material, Long> materialDao;
+    private Dao<Contact, Long> contactDao;
 
     @Inject
     public DatabaseHelper(Context context) {
@@ -61,35 +62,35 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         }
     }
 
-    public Dao<Event, Integer> getEventDao() throws SQLException {
+    public Dao<Event, Long> getEventDao() throws SQLException {
         if (eventDao == null) {
             eventDao = getDao(Event.class);
         }
         return eventDao;
     }
 
-    public Dao<Article, Integer> getArticleDao() throws SQLException {
+    public Dao<Article, Long> getArticleDao() throws SQLException {
         if (articleDao == null) {
             articleDao = getDao(Article.class);
         }
         return articleDao;
     }
 
-    public Dao<Sponsor, Integer> getSponsorDao() throws SQLException {
+    public Dao<Sponsor, Long> getSponsorDao() throws SQLException {
         if (sponsorDao == null) {
             sponsorDao = getDao(Sponsor.class);
         }
         return sponsorDao;
     }
 
-    public Dao<Material, Integer> getMaterialDao() throws SQLException {
+    public Dao<Material, Long> getMaterialDao() throws SQLException {
         if (materialDao == null) {
             materialDao = getDao(Material.class);
         }
         return materialDao;
     }
 
-    public Dao<Contact, Integer> getContactDao() throws SQLException {
+    public Dao<Contact, Long> getContactDao() throws SQLException {
         if (contactDao == null) {
             contactDao = getDao(Contact.class);
         }
